requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
discord==1.0.1
discord.py==1.5.0
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.2
idna==2.10
multidict==4.7.6
numpy==1.19.3
pandas==1.1.4
praw==7.1.0
prawcore==1.5.0
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
update-checker==0.18.0
urllib3==1.25.11
websocket-client==0.57.0
yarl==1.5.1
youtube_dl
youtube-search-python==1.3.1
google-api-python-client
python-dotenv
